animations

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { View, FlatList,Text } from 'react-native';
+import { View, FlatList,Text , Alert} from 'react-native';
 import { ListItem } from "react-native-elements";
 import { connect} from 'react-redux';
 import { baseURL } from '../shared/baseURL';
@@ -35,7 +35,25 @@ class Favorite extends Component {
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite',
+                            'Are you sure you wish to delete this favorite Dish ' + item.name +'?',
+                            [
+                                {
+                                    text:'Cancel',
+                                    onPress: () => console.log(item.name + 'Not deleted'),
+                                    style: 'cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id),
+                                }
+                            ],
+                            { cancelable: false}
+                        )
+                        
+                    }
                 }
             ]
